--- airbyte-webapp/src/components/Breadcrumbs/index.tsx
@@ -1,3 +1,0 @@
-import Breadcrumbs from "./Breadcrumbs";
-
-export default Breadcrumbs;
